requirements.txt
python-twitter
fbchat
pushbullet.py
